XAudio2 bugs resolved
Buffers given to XAudio2 are not copied, so are not allowed to change or move. Adjusted feedback functions to account for this.
FFmpeg audio stream functions could not correctly place data in the middle of an array, now fixed
Tuned the RGAdaptive filter a bit more,  added an upper cap to the amplification to avoid clipping bursts